CPM-629: Move classes and fix reviews
