+++ src/Akeneo/Platform/Bundle/InstallerBundle/Resources/config/subscribers.yml
@@ -4,3 +4,10 @@ services:
       - '@database_connection'
     tags:
       - { name: 'kernel.event_subscriber' }
+
+  Akeneo\Platform\Bundle\InstallerBundle\Event\Subscriber\MarkZddMigrationsAsMigratedSubscriber:
+    arguments:
+      - '@database_connection'
+      - !tagged_iterator { tag: 'akeneo.pim.zdd_migration' }
+    tags:
+      - { name: kernel.event_subscriber }
